refactor: Organize into packages to meet requirements
