+++ src/main/java/marshmallow/Marshmallow.java
@@ -1,4 +1,14 @@
+package marshmallow;
+
 import java.util.Scanner;
+
+import marshmallow.commands.Command;
+import marshmallow.parser.Parser;
+import marshmallow.storage.textfilestorage.TextFileStorage;
+import marshmallow.tasks.Task;
+import marshmallow.tasks.TaskFactory;
+import marshmallow.tasks.TaskList;
+import marshmallow.ui.Ui;
 
 public class Marshmallow {
     private TextFileStorage<Task, TaskFactory> storage;
