+++ src/main/java/marshmallow/MarshmallowException.java
@@ -1,3 +1,5 @@
+package marshmallow;
+
 public class MarshmallowException extends RuntimeException {
     public MarshmallowException(String message) {
         super(message);
