+++ src/main/java/marshmallow/commands/Command.java
@@ -1,3 +1,8 @@
+package marshmallow.commands;
+
+import marshmallow.tasks.TaskList;
+import marshmallow.ui.Ui;
+
 public abstract class Command {
     public boolean isExit;
