+++ src/main/java/marshmallow/commands/ExitCommand.java
@@ -1,3 +1,8 @@
+package marshmallow.commands;
+
+import marshmallow.tasks.TaskList;
+import marshmallow.ui.Ui;
+
 public class ExitCommand extends Command {
     public ExitCommand() {
         this.isExit = true;
