+++ src/main/java/marshmallow/commands/ListAddDeadlineCommand.java
@@ -1,3 +1,8 @@
+package marshmallow.commands;
+
+import marshmallow.MarshmallowException;
+import marshmallow.tasks.Deadline;
+
 public class ListAddDeadlineCommand extends ListAddTaskCommand {
     public ListAddDeadlineCommand(String description, String endTime) {
         if (endTime == null) {
