+++ src/main/java/marshmallow/commands/ListAddTaskCommand.java
@@ -1,3 +1,9 @@
+package marshmallow.commands;
+
+import marshmallow.tasks.Task;
+import marshmallow.tasks.TaskList;
+import marshmallow.ui.Ui;
+
 public abstract class ListAddTaskCommand extends Command {
     protected Task task;
