+++ src/main/java/marshmallow/commands/ListDeleteTaskCommand.java
@@ -1,3 +1,9 @@
+package marshmallow.commands;
+
+import marshmallow.tasks.Task;
+import marshmallow.tasks.TaskList;
+import marshmallow.ui.Ui;
+
 public class ListDeleteTaskCommand extends Command {
     private int idx;
