+++ src/main/java/marshmallow/storage/textfilestorage/TextFileStorageConstructable.java
@@ -1,3 +1,5 @@
+package marshmallow.storage.textfilestorage;
+
 public interface TextFileStorageConstructable<T> {
     public T constructFromSaveString(String data);
 }
