+++ src/main/java/marshmallow/storage/textfilestorage/TextFileStorageSaveable.java
@@ -1,3 +1,5 @@
+package marshmallow.storage.textfilestorage;
+
 public interface TextFileStorageSaveable {
     public String toSaveString();
 }
