+++ src/main/java/marshmallow/tasks/Event.java
@@ -1,23 +1,27 @@
+package marshmallow.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
+    private LocalDateTime startTime;
     private LocalDateTime endTime;
     private static DateTimeFormatter dtfParse = DateTimeFormatter.ofPattern("dd-MM-yyyy HHmm");
     private static DateTimeFormatter dtfDisplay = DateTimeFormatter.ofPattern("MMM dd yyyy HHmm");
 
-    public Deadline(String description, String endTimeString) {
+    public Event(String description, String startTimeString, String endTimeString) {
         super(description);
+        this.startTime = LocalDateTime.parse(startTimeString, dtfParse);
         this.endTime = LocalDateTime.parse(endTimeString, dtfParse);
     }
 
     @Override
     public String toSaveString() {
-        return String.format("D | %s | %s", super.toSaveString(), endTime.format(dtfParse));
+        return String.format("E | %s | %s | %s", super.toSaveString(), startTime.format(dtfParse), endTime.format(dtfParse));
     }
 
     @Override
     public String toString() {
-        return String.format("[D]%s (by: %s)", super.toString(), endTime.format(dtfDisplay));
+        return String.format("[E]%s (from: %s to: %s)", super.toString(), startTime.format(dtfDisplay), endTime.format(dtfDisplay));
     }
 }
